# apps/web/apps/web-antd/src/api/generated/infra.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RustSet Infra API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/infra/cloud-zone/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/cloud-zone/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudZoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/cloud-zone/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCloudZoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/cloud-platform/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/cloud-platform/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudPlatformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/cloud-platform/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCloudPlatformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/cloud-provider-config/page": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "pageNo",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": null
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_Page_for_AnyValue"
                }
              }
            }
          }
        }
      }
    },
...
    "/infra/cloud-provider-config/create": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCloudProviderConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_string"
                }
              }
            }
          }
        }
      }
    },
    "/infra/cloud-provider-config/update": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCloudProviderConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_for_null"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_for_Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "$ref": "#/components/schemas/Page_for_AnyValue"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "ApiResponse_for_null": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "null"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
      "ApiResponse_for_string": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "uint16",
            "maximum": 65535,
            "minimum": 0
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "data",
          "message"
        ]
      },
...
      "CreateCloudPlatformRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "platformCode": {
            "type": "string"
          },
          "platformName": {
            "type": "string"
          },
          "zoneId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "required": [
          "zoneId",
          "platformName",
          "platformCode"
        ]
      },
      "CreateCloudProviderConfigRequest": {
        "type": "object",
        "properties": {
          "accessKeyId": {
            "type": "string"
          },
          "accessKeySecret": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "availableZones": {
            "type": [
              "string",
              "null"
            ]
          },
          "platformId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "provider": {
            "type": "string"
          },
          "regionId": {
            "type": "string"
          },
          "regionName": {
            "type": "string"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "default": "active"
          },
          "zoneId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "required": [
          "provider",
          "regionId",
          "regionName",
          "accountName",
          "accessKeyId",
          "accessKeySecret"
        ]
      },
      "CreateCloudZoneRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "zoneCode": {
            "type": "string"
          },
          "zoneName": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "zoneName",
          "zoneCode"
        ]
      },
...
      "Page_for_AnyValue": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": [
          "list",
          "total"
        ]
      },
...
      "UpdateCloudPlatformRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "platformCode": {
            "type": "string"
          },
          "platformName": {
            "type": "string"
          },
          "zoneId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "zoneId",
          "platformName",
          "platformCode"
        ]
      },
      "UpdateCloudProviderConfigRequest": {
        "type": "object",
        "properties": {
          "accessKeyId": {
            "type": "string"
          },
          "accessKeySecret": {
            "type": "string"
          },
          "accountName": {
            "type": "string"
          },
          "availableZones": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "platformId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "provider": {
            "type": "string"
          },
          "regionId": {
            "type": "string"
          },
          "regionName": {
            "type": "string"
          },
          "remarks": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string",
            "default": "active"
          },
          "zoneId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "provider",
          "regionId",
          "regionName",
          "accountName",
          "accessKeyId",
          "accessKeySecret"
        ]
      },
      "UpdateCloudZoneRequest": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "zoneCode": {
            "type": "string"
          },
          "zoneName": {
            "type": "string"
          }
        },
        "additionalProperties": false,
        "required": [
          "id",
          "zoneName",
          "zoneCode"
        ]
      },
...
    }
  }
}